Extract shared SMTP banner parsing to internal/protocol/smtp
Moves parseSMTPBanner and readBanner into a shared package at
internal/protocol/smtp, eliminating duplication between the discover
fingerprinter and enumerate modules. Also switches readBanner to
bufio-based line reading to handle TCP fragmentation reliably.
