+++ internal/enumerate/smtp/helpers.go
@@ -2,11 +2,9 @@ package smtp
 
 import (
 	// Standard
-	"bytes"
 	"context"
 	"crypto/tls"
 	"fmt"
-	"io"
 	"net"
 	netsmtp "net/smtp"
 	"strings"
@@ -19,90 +17,26 @@ import (
 
 func tryTCPConnection(ctx context.Context, target string) (net.Conn, error) {
 	dialer := net.Dialer{}
-	conn, err := dialer.DialContext(ctx, "tcp", target)
-	if err != nil {
-		return nil, err
-	}
-	return conn, nil
+	return dialer.DialContext(ctx, "tcp", target)
 }
 
-// readBanner reads the SMTP 220 banner by peeking at the connection data
-// without consuming it, so net/smtp.NewClient can still read it.
-// Returns a wrapped conn that replays the banner bytes.
-func readBanner(conn net.Conn) (net.Conn, string, error) {
-	buf := make([]byte, 4096)
-	n, err := conn.Read(buf)
-	if err != nil {
-		return conn, "", err
+func tryTLSConnection(target string, hostname string) (net.Conn, error) {
+	dialer := net.Dialer{}
+	tlsConfig := &tls.Config{
+		ServerName:         hostname,
+		InsecureSkipVerify: true,
 	}
-	data := buf[:n]
-
-	// Extract banner from the first line
-	banner := ""
-	lines := strings.Split(string(data), "\r\n")
-	if len(lines) > 0 && strings.HasPrefix(lines[0], "220") {
-		banner = strings.TrimSpace(lines[0])
-	}
-
-	// Wrap the connection so NewClient can re-read the banner
-	wrapped := &replayConn{
-		Conn:   conn,
-		reader: io.MultiReader(bytes.NewReader(data), conn),
-	}
-	return wrapped, banner, nil
+	return tls.DialWithDialer(&dialer, "tcp", target, tlsConfig)
 }
 
-// replayConn wraps a net.Conn, replaying buffered data before reading from the real connection.
-type replayConn struct {
-	net.Conn
-	reader io.Reader
-}
-
-func (c *replayConn) Read(p []byte) (int, error) {
-	return c.reader.Read(p)
-}
-
-// parseSMTPBanner extracts server name and software info from a 220 banner.
-func parseSMTPBanner(banner string) (serverName, softwareName, softwareVersion string) {
-	line := banner
-	if strings.HasPrefix(line, "220-") {
-		line = line[4:]
-	} else if strings.HasPrefix(line, "220 ") {
-		line = line[4:]
-	} else {
-		return
-	}
-
-	parts := strings.Fields(line)
-	if len(parts) == 0 {
-		return
-	}
-
-	serverName = parts[0]
-
-	idx := 1
-	for idx < len(parts) {
-		upper := strings.ToUpper(parts[idx])
-		if upper == "ESMTP" || upper == "SMTP" {
-			idx++
-			break
+func parseAuthMethods(methods []string) []smtp.AuthCommand {
+	var result []smtp.AuthCommand
+	for _, method := range methods {
+		if auth, ok := authCommands[strings.ToUpper(method)]; ok {
+			result = append(result, auth)
 		}
-		idx++
 	}
-
-	if idx < len(parts) {
-		remaining := parts[idx:]
-		for i, token := range remaining {
-			if len(token) > 0 && token[0] >= '0' && token[0] <= '9' {
-				softwareName = strings.Join(remaining[:i], " ")
-				softwareVersion = strings.Join(remaining[i:], " ")
-				return
-			}
-		}
-		softwareName = strings.Join(remaining, " ")
-	}
-
-	return
+	return result
 }
 
 // collectExtensions checks for common SMTP extensions via the client's Extension method.
@@ -123,25 +57,6 @@ func collectExtensions(c *netsmtp.Client) []string {
 		}
 	}
 	return found
-}
-
-func tryTLSConnection(target string, hostname string) (net.Conn, error) {
-	dialer := net.Dialer{}
-	tlsConfig := &tls.Config{
-		ServerName:         hostname,
-		InsecureSkipVerify: true,
-	}
-	return tls.DialWithDialer(&dialer, "tcp", target, tlsConfig)
-}
-
-func parseAuthMethods(methods []string) []smtp.AuthCommand {
-	var result []smtp.AuthCommand
-	for _, method := range methods {
-		if auth, ok := authCommands[strings.ToUpper(method)]; ok {
-			result = append(result, auth)
-		}
-	}
-	return result
 }
 
 func testUnauthenticatedEmail(ctx context.Context, c *netsmtp.Client, hostname string) bool {
